backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
